Refactor Variable into an interface in a separate package
I had a number of bugs related to how the Variable type was being
created where, as we added new fields, some of the old places using
that variable were not providing reasonable defaults for those new
fields. Moving it into a separate package, turning it into an
interface, and making the implementation private allows us to precisely
control “constructor” usage and ensure all necessary defaults are set.
It also provides a cleaner API for working with the Variable class.