change listSlice to sliceList
In the documentation this function alias is referred to as "sliceList".
When trying to use it I kept getting

```
function "sliceList" not defined
```

After going back and forth in the docs, double-checking my typing,
copy-pasting from the docs, updating to the latest release, I eventually
decided to dive into the code. Thankfully it was pretty obvious what
went wrong here. I even went back to my template to verify the
`listSlice` worked and it did! Hooray! Of course next release I'll need
to change my `listSlice` to `sliceList`, but at least everything will be
aligned now :)

Signed-off-by: Taylor Silva <dev@taydev.net>